(PC-17818) chore(venue): move atoms to component folder
