+++ src/features/venue/components/VenueType.native.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 
 import { VenueTypeCodeKey } from 'api/gen'
+import { VenueType } from 'features/venue/components/VenueType'
 import { parseType } from 'libs/parsers'
 import { render } from 'tests/utils'
-
-import { VenueType } from '../VenueType'
 
 const typeLabel = parseType(VenueTypeCodeKey.MOVIE)
