+++ src/features/venue/components/VenueType.web.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 
 import { VenueTypeCodeKey } from 'api/gen'
+import { VenueType } from 'features/venue/components/VenueType'
 import { parseType } from 'libs/parsers'
 import { render } from 'tests/utils/web'
-
-import { VenueType } from '../VenueType'
 
 const typeLabel = parseType(VenueTypeCodeKey.MOVIE)
